refactor(auth): move authentication logic into dedicated Auth class
refactor(auth): move authentication logic into dedicated Auth class

- Create dedicated Auth class for authentication handling
- Move authentication methods from API class to Auth class
- Update namespace and class references for Auth
- Add comprehensive inline documentation for Auth class
- Clean up API class authentication logic
- Remove redundant code in API class
- Use Auth service class for all auth operations

BREAKING CHANGES:
- Authentication methods moved to Auth class
- Token validation now uses Auth::validateToken()
- Login handling moved to Auth::login()

Closes #124
