--- api/Auth/Auth.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types=1);
-namespace Glueful\Auth;
-
-class Auth
-{
-    // Authentication logic will go here
-}
